script base
To use: add to mu-plugins folder. go to the site and set keyhashing as a GET key. 

NOTE: this can ONLY be run once!!!!! one ran, the keys are hashed and if ran again, you are rehashing hashes. Dear oh dear. 

TODO: build into a plugin and build in feature to only allow to be run once. eg do a check to see if the key is already hashed or not.